pyproject.toml
[tool.poetry]
name = "vpragati"
version = "0.1.0"
description = ""
authors = ["HarshPatel285 <harshpatel@glib.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.2"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
rules = "^3.3"
crispy-bootstrap5 = "^0.7"
phonenumbers = "^8.13.15"
django-phonenumber-field = "^7.1.0"
django-bootstrap-datepicker-plus = "^5.0.3"
django-tables2 = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
